build and update controllers and add auth

+++ controllers/clothingitems.js
@@ -31,38 +31,57 @@ const getItems = (req, res) => {
     });
 };
 
-const updateItem = (req, res) => {
-  const { itemId } = req.params;
-  const { imageUrl } = req.body;
+const updateItem = async (req, res) => {
+  try {
+    const { itemId } = req.params;
+    const { imageUrl } = req.body;
 
-  ClothingItem.findByIdAndUpdate(itemId, { $set: { imageUrl } })
-    .orFail()
-    .then((item) => res.status(200).send({ data: item }))
-    .catch((err) => {
-      if (err.name === "DocumentNotFoundError") {
-        return res.status(404).send({ message: err.message });
-      }
-      if (err.name === "CastError") {
-        return res.status(400).send({ message: err.message });
-      }
-      return res.status(500).send({ message: err.message });
-    });
+    const item = await ClothingItem.findById(itemId);
+    if (!item) {
+      return res.status(404).send({ message: "Item not found" });
+    }
+    if (item.owner.toString() !== req.user._id.toString()) {
+      return res
+        .status(403)
+        .send({ message: "You don't have permission to update this item" });
+    }
+
+    const updatedItem = await ClothingItem.findByIdAndUpdate(
+      itemId,
+      { $set: { imageUrl } },
+      { new: true }
+    );
+    res.status(200).send({ data: updatedItem });
+  } catch (err) {
+    if (err.name === "CastError") {
+      return res.status(400).send({ message: err.message });
+    }
+    return res.status(500).send({ message: err.message });
+  }
 };
 
-const deleteItem = (req, res) => {
-  const { itemId } = req.params;
-  ClothingItem.findByIdAndDelete(itemId)
-    .orFail()
-    .then((item) => res.status(200).send({ data: item }))
-    .catch((err) => {
-      if (err.name === "DocumentNotFoundError") {
-        return res.status(404).send({ message: err.message });
-      }
-      if (err.name === "CastError") {
-        return res.status(400).send({ message: err.message });
-      }
-      return res.status(500).send({ message: err.message });
-    });
+const deleteItem = async (req, res) => {
+  try {
+    const { itemId } = req.params;
+
+    const item = await ClothingItem.findById(itemId);
+    if (!item) {
+      return res.status(404).send({ message: "Item not found" });
+    }
+    if (item.owner.toString() !== req.user._id.toString()) {
+      return res
+        .status(403)
+        .send({ message: "You don't have permission to delete this item" });
+    }
+
+    const deletedItem = await ClothingItem.findByIdAndDelete(itemId);
+    res.status(200).send({ data: deletedItem });
+  } catch (err) {
+    if (err.name === "CastError") {
+      return res.status(400).send({ message: err.message });
+    }
+    return res.status(500).send({ message: err.message });
+  }
 };
 
 module.exports = { createItem, getItems, updateItem, deleteItem };
